Add attrsEqual and marksEqual utilities to core
Export attrsEqual and marksEqual from @tiptap/core utilities.
Add unit tests for both, remove duplicated implementations in the
markdown package, and add a changeset
