+++ packages/core/src/utilities/index.ts
@@ -1,3 +1,4 @@
+export * from './attrsEqual.js'
 export * from './callOrReturn.js'
 export * from './canInsertNode.js'
 export * from './createStyleTag.js'
@@ -21,6 +22,7 @@ export * from './isSafari.js'
 export * from './isString.js'
 export * from './markdown/index.js'
 export * as markdown from './markdown/index.js'
+export * from './marksEqual.js'
 export * from './mergeAttributes.js'
 export * from './mergeDeep.js'
 export * from './minMax.js'
